add crud users, products
add something dumb

+++ week-3/order-service/services/user/entity/user.go
@@ -6,7 +6,7 @@ type User struct {
 	Id        int        `json:"id"`
 	Username  string     `json:"username"`
 	Password  string     `json:"-"` // sensitive field, should not send to user
-	Balance   float64    `json:"balance"`
+	Balance   float32    `json:"balance"`
 	CreatedAt time.Time  `json:"created_at"`
 	UpdatedAt *time.Time `json:"updated_at"`
 }
